streams: split generic plumbing into streamtools/ and extract decompressor base
Two structural cleanups to the Phase-2 stream layer, no behavior change:

1. New streams/streamtools/ subpackage holds the codec- and archivey-agnostic
   plumbing — binaryio (adapt/classify/coerce arbitrary BinaryIO) and slice
   (SlicingStream). It imports nothing from archivey's domain (the lone tie,
   binaryio's logger, now uses logging.getLogger("archivey.streams")), so it
   could be lifted out as a standalone library. ArchiveStream and the codec
   backends stay behind: they speak ArchiveyError and are archivey-specific.

2. decompress.py conflated the codec-agnostic seekable base with concrete
   backends. Extract DecompressorStream / _SegmentedDecompressorStream /
   SeekPoint into decompressor_stream.py; decompress.py now holds only the
   concrete zlib/deflate + Brotli backends. xz/lzip import the base from the
   new module, so "import the base" no longer drags in sibling backends.

Import sites, the streamtools __init__ public surface, and ARCHITECTURE.md §1
/ §2.5 updated. Renames preserve history. ruff/ty/pyrefly clean; 371 passed.
